requirements.txt
argparse~=1.4.0
PyYAML~=6.0
torch~=1.10.0
numpy~=1.21.3
pandas~=1.3.4
transformers~=4.14.1
matplotlib~=3.4.3
sklearn~=0.0
scikit-learn~=1.0.1
tqdm~=4.62.3
langdetect~=1.0.9
Gutenberg~=0.4.0
streamlit
tweepy
nltk
pandas
transformers
torch
numpy
emoji
pillow